refactored auth code :)
